解决 clipy warning

+++ crates/shadowsocks/src/vless/mux/server.rs
@@ -169,7 +169,7 @@ async fn read_data(
     if len > 0 {
         session
             .read_cmd_sender()
-            .send(SessionReadCmd::Read(len as usize))
+            .send(SessionReadCmd::Read(len))
             .await
             .map_err(|e| new_error(format!("read_data: send read cmd fail {}", e)))?;
         let _ = read_done_receiver.recv().await;
